Refactored online request functionality.
